Close cast ownership and physical organization (#592)
Resolve the semantic and physical closeout of #589 with reviewed deltas
instead of blind baseline regeneration.

session_tests.rs stood at 95705 lines against a 95703 ceiling. Rather
than raise it, the 1015-line player spell-hit source authority suite
moves into the bounded session/tests/player_spell_hit_source.rs with its
registrations unchanged, leaving 94691 lines; the ceiling is tightened to
that validated figure. handlers/spell.rs drops from 6737 to 6282 as the
cast responsibility moves into player_cast, and session/mod.rs from 76107
to 75840; both ceilings are tightened. handlers/spell.rs stays above the
2000-line budget, so its separation remains open under #584 C4 and is not
declared closed by this reduction.

Three files exceed their ceiling by an inseparable delta and are recorded
as expressly justified rather than split: main_tests.rs +2 for naming the
new typed SpellCastData field in two fixtures, session_state.rs +4 for
the binder consumer adapting to fallible canonical allocation, and
directory.rs +16 for the single PlayerRegistry publication entry point
that mirrors its creature equivalent.

The ownership syntax baseline is regenerated from a checker built in this
checkout, and its delta reviewed item by item: one new Runtime impl for
WorldSession, the relocated cast methods, one new SessionCommand variant
with its payload and derive, two registry accesses in the publication
adapter, and one bridge fingerprint change from the edited Go path. No
new fully public surface; the only widened visibilities are private to
pub(crate) or pub(in crate::session), the minimum for the split.
can_request_represented_spell_cast_like_cpp retires with no replacement:
the admission window now belongs to player_cast::request.

The runtime ownership ledger records the measured growth with a dated
review and registers the new SessionCommand variant in the map runtime
and visibility delivery group.

Executed in this checkout:
- session-ownership-check check --syntax-only: PASS (55 impl owners,
  3761 exact associated items, 40 SessionCommand variants, 594 registry
  rows; exact persistence inventory deferred to push and cron).
- check_architecture.py check: PASS (physical ratchet 1067 files,
  100 legacy ceilings, 178 cohesion reviews).
- check_architecture.py self-test: PASS (20 fixtures).
- cargo check -p wow-world --lib --all-targets: passes, 35 warnings,
  unchanged from before these edits.
