+++ tools/architecture/physical-file-policy.json
@@ -62,9 +62,9 @@
     },
     {
       "path": "crates/world-server/src/main_tests.rs",
-      "observed_lines": 13040,
-      "ceiling_lines": 13040,
-      "split": "Separate configuration, composition, shutdown, session supervision and runtime timing suites with narrow fixture builders.",
+      "observed_lines": 13042,
+      "ceiling_lines": 13042,
+      "split": "Separate configuration, composition, shutdown, session supervision and runtime timing suites with narrow fixture builders. #589 adds two inseparable lines: SpellStartPkt and SpellGoPkt now carry the typed SpellCastData payload, so both creature-plan fixtures name the field explicitly instead of relying on the former hardcoded empty sections. No new scenario is added here.",
       "review_checkpoint": "584:C4"
     },
     {
@@ -468,9 +468,9 @@
     },
     {
       "path": "crates/wow-world/src/handlers/character/session_state.rs",
-      "observed_lines": 2806,
-      "ceiling_lines": 2806,
-      "split": "Separate character adapters by complete operation; preserve registration/publication order. #585 adds ten lines so both failed-login consumers retain identity when canonical retirement fails, rather than discarding the returned outcome. This bounded safety integration does not retire the remaining login/worldport module debt in #584.",
+      "observed_lines": 2810,
+      "ceiling_lines": 2810,
+      "split": "Separate character adapters by complete operation; preserve registration/publication order. #585 adds ten lines so both failed-login consumers retain identity when canonical retirement fails, rather than discarding the returned outcome. This bounded safety integration does not retire the remaining login/worldport module debt in #584. #589 adds four inseparable lines: the binder consumer adapts to fallible canonical cast-identity allocation and returns when the admitted Map cannot allocate, instead of publishing an empty CastID. No new responsibility enters this file.",
       "review_checkpoint": "584:C4"
     },
     {
@@ -573,9 +573,9 @@
     },
     {
       "path": "crates/wow-world/src/handlers/spell.rs",
-      "observed_lines": 6737,
-      "ceiling_lines": 6737,
-      "split": "Separate spell admission/cast operations by complete operation into narrow adapters, owner capabilities and scenario tests; preserve opcode registration and publication order.",
+      "observed_lines": 6282,
+      "ceiling_lines": 6282,
+      "split": "Separate spell admission/cast operations by complete operation into narrow adapters, owner capabilities and scenario tests; preserve opcode registration and publication order. #589 moves normal request admission, preparation, identity, checks, power and publication into the private wow-world::player_cast application and its Session adapters, reducing this handler by 455 lines; the ceiling is tightened to the validated 6282. The remaining file still owns the broader spell handler responsibility and stays above the 2000-line budget, so its separation stays in #584 C4 rather than being declared closed by this reduction.",
       "review_checkpoint": "584:C4"
     },
     {
@@ -622,23 +622,23 @@
     },
     {
       "path": "crates/wow-world/src/session/directory.rs",
-      "observed_lines": 2335,
-      "ceiling_lines": 2335,
-      "split": "Separate directory registration/lifetime, recipient queries and fanout; keep it derived rather than gameplay authority.",
+      "observed_lines": 2351,
+      "ceiling_lines": 2351,
+      "split": "Separate directory registration/lifetime, recipient queries and fanout; keep it derived rather than gameplay authority. #589 adds sixteen inseparable lines: one PlayerRegistry publication entry point for represented Player spell frames over the existing bounded durable rail, alongside the creature equivalent it mirrors. Relocating a single registry accessor would fragment the registry's own responsibility.",
       "review_checkpoint": "584:C4"
     },
     {
       "path": "crates/wow-world/src/session/mod.rs",
-      "observed_lines": 76107,
-      "ceiling_lines": 76107,
-      "split": "Retire gameplay orchestration/catalog access by complete operations. #585 adds the task-owned finalization ledger and identity-hydration guard. PR #586 review adds 16 inseparable lines to retain the native far-transfer destination before an interruptible realm writer fence in both immediate and delayed paths. Existing update methods become async in the same Session task; no new field, task, lock or state mirror. New failure/cancellation tests remain in bounded files. This reviewed growth is not physical closeout: the remaining C4 root split stays in #584.",
+      "observed_lines": 75840,
+      "ceiling_lines": 75840,
+      "split": "Retire gameplay orchestration/catalog access by complete operations. #585 adds the task-owned finalization ledger and identity-hydration guard. PR #586 review adds 16 inseparable lines to retain the native far-transfer destination before an interruptible realm writer fence in both immediate and delayed paths. #589 moves the represented Player cast publication out to session/player_cast/, a net reduction of 267 lines; the ceiling is tightened to the validated 75840. Existing update methods become async in the same Session task; no new field, task, lock or state mirror. New failure/cancellation tests remain in bounded files. This reviewed growth is not physical closeout: the remaining C4 root split stays in #584.",
       "review_checkpoint": "584:C4"
     },
     {
       "path": "crates/wow-world/src/session_tests.rs",
-      "observed_lines": 96722,
-      "ceiling_lines": 95703,
-      "split": "Move lifecycle, transfer, spells, quests, inventory, social and runtime scenario suites to their corresponding owners; preserve each test registration. Above e8a3fb1f +7 lines extend existing exact teleport state through active/detached/replaced ownership and reject stale completion without changing the replacement. New scenarios stay in bounded files. C4 split exit remains.",
+      "observed_lines": 94691,
+      "ceiling_lines": 94691,
+      "split": "Move lifecycle, transfer, spells, quests, inventory, social and runtime scenario suites to their corresponding owners; preserve each test registration. #589 retires the 1015-line player spell-hit source authority suite into session/tests/player_spell_hit_source.rs and registers the bounded player_cast_lifecycle module, a net reduction of 2031 lines below the previous 95703 ceiling; the ceiling is tightened to the validated 94691. New scenarios stay in bounded files. C4 split exit remains.",
       "review_checkpoint": "584:C4"
     },
     {
